% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_tools.R
\name{build_GOTM}
\alias{build_GOTM}
\title{Function to compile GOTM-FABM}
\usage{
build_GOTM(build_dir, src_dir, fabm_file)
}
\arguments{
\item{build_dir}{build directory}

\item{src_dir}{source directory. should be the FABM directory within the GOTM source directory}

\item{fabm_file}{forthran source file for FABM created by gen_fabm_code()}
}
\description{
This function compiles GOTM-FABM with FABM source code generated by gen_fabm_code().
It assumes that you allready have run cmake within the build directory
}
\examples{
\dontrun{
library(readODS)
library(rodeoFABM)

# copy example ods file
example_model <- system.file("extdata/simple_model.ods", package= 'rodeoFABM')
dir.create('example') # Create example folder
file.copy(from = example_model, to = 'example',recursive = TRUE)
setwd('example') # Change working directory to example folder

# read in example ods file
odf_file <- "simple_model.ods"
vars <- read_ods(odf_file,1)
pars <- read_ods(odf_file,2)
funs <- read_ods(odf_file,3)
pros <- read_ods(odf_file,4)
stoi <- read_ods(odf_file,5)

# generate fabm code
gen_fabm_code(vars,pars,funs,pros,stoi,"simple_model.f90",diags = TRUE)
# build GOTM
build_GOTM(build_dir = "../build",fabm_file = "simple_model.f90",
           src_dir = "../gotm/extern/fabm/src/models/tuddhyb/rodeo")
}

}
\author{
Johannes Feldbauer
}
\keyword{FABM,}
\keyword{GOTM,}
\keyword{compile}
