% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_model.R
\name{document_model}
\alias{document_model}
\title{Function to create LaTeX documentation of a model}
\usage{
document_model(vars, pars, pros, funs, stoi, landscape = TRUE, tex = "tex")
}
\arguments{
\item{vars}{data.frame containing the state variables}

\item{pars}{data.frame containing the parameters}

\item{pros}{data.frame containing the processes}

\item{funs}{data.frame containing the functions}

\item{stoi}{data.frame containing the stoichiometry}

\item{landscape}{boolean, should the process expression table be in landscape?}

\item{tex}{name of the column containing the LaTeX expressions to use as symbols}
}
\description{
This function creates LaTeX tables from the tables used to generate the FABM source code.
It will generate a table with the statevariables, a table with the model parameter, a table
with tzhe process names and descriptions, a table with the process expressions, and a table
with the stoichiometry table.
}
\examples{
\dontrun{
library(readODS)
library(rodeoFABM)

# copy example ods file
example_model <- system.file("extdata/simple_model.ods", package= 'rodeoFABM')
dir.create('example') # Create example folder
file.copy(from = example_model, to = 'example',recursive = TRUE)
setwd('example') # Change working directory to example folder

# read in example ods file
odf_file <- "simple_model.ods"
vars <- read_ods(odf_file,1)
pars <- read_ods(odf_file,2)
funs <- read_ods(odf_file,3)
pros <- read_ods(odf_file,4)
stoi <- read_ods(odf_file,5)

# generate documentation
document_model(vars,pars,pros,funs,stoi)
}

}
\author{
Johannes Feldbauer
}
\keyword{FABM,}
\keyword{GOTM,}
\keyword{LaTeX}
\keyword{document,}
