% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypsograph}
\alias{hypsograph}
\title{Generic hypsographic curve.}
\format{A text file (.dat) with two columns
\describe{
  \item{level}{Depth below water surface in m}
  \item{Area}{Surface area at depth in m^2}

}}
\usage{
data(hypsograph)
}
\description{
Hypsographic curve for example lake following a second order polynom
}
\keyword{datasets}
