% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo_file}
\alias{meteo_file}
\title{Example meteorological forcing data for GOTM.}
\format{A text file (.dat) with nine columns
\describe{
  \item{datetime}{Day of the observation}
  \item{Uwind}{North-South wind speed in m/s}
  \item{Vwind}{East-West wind speed in m/s}
  \item{pressure}{Surface level barometric pressure in pascal}
  \item{airTemp}{Air temperature in degree celsius}
  \item{relHum}{Relative humidity in percent}
  \item{cloudCover}{Cloud cover as decimal fraction}
  \item{short}{Shortwave radiation downwelling in W/m^2}
  \item{rain}{Daily average precipitation in m/s}
}}
\source{
\url{https://opendata.dwd.de/climate_environment/CDC/}
}
\usage{
data(meteo_file)
}
\description{
Meteorologic forcing data for running the GOTM-FABM test example. The data was averaged
from different DWD (German Weather Service) stations for the period of 2009 to 2012.
}
\keyword{datasets}
