% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_tools.R
\name{plot_var}
\alias{plot_var}
\title{Plot parameter from a netcdf file created by GOTM}
\usage{
plot_var(
  file = "output.nc",
  var = "temp",
  reference = "surface",
  z_out = NULL,
  t_out = NULL,
  res = 0.25,
  add = FALSE,
  col = 1,
  main = TRUE,
  colp = mycol(100)
)
}
\arguments{
\item{file}{Path to the netcdf file}

\item{var}{Name of the variable to get}

\item{reference}{Either "surface" or "bottom". The reference point for the depths sepified in
'z_out'}

\item{z_out}{Vector of depths from which to get the variable. If no depths are selected
(NULL, the default) all depths from the lowest to the highest depth in a equal distance of
'res'}

\item{t_out}{Vector of times (as POSIX) for which the variable should be fetched. If no times
are selected (NULL, the default) all available times from the netcdf file are returned}

\item{res}{Spatial resolution for depths, if no depths are explicitly specified by 'z_out'}

\item{add}{If only one depth is selected, add to an existing plot}

\item{col}{If only one depth is selected, color of the line to plot}

\item{main}{Title of the plot}

\item{colp}{Color palette to use for the 2D image}
}
\description{
This function plots variables for given depths and times from a netcdf file created
by running GOTM-FABM
}
\examples{
\dontrun{
plot_var(file = "output.nc", var = "temp")
}

}
\author{
Johannes Feldbauer
}
\keyword{FABM,}
\keyword{GOTM,}
\keyword{get}
\keyword{variable}
