% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_FABM.R
\name{gen_fabm_code}
\alias{gen_fabm_code}
\title{Function to generate FABM code}
\usage{
gen_fabm_code(
  vars,
  pars,
  funs,
  pros,
  stoi,
  file_name = "model.f90",
  diags = TRUE
)
}
\arguments{
\item{vars}{data.frame containing the state variables}

\item{pars}{data.frame containing the parameters}

\item{funs}{data.frame containing the functions}

\item{pros}{data.frame containing the processes}

\item{stoi}{data.frame containing the stoichiometry}
}
\description{
This function creates fortran90 FABM code from data.frames.
}
\keyword{FABM}
