% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_names_FABM}
\alias{std_names_FABM}
\title{standard names for dependencie variables in FABM.}
\format{A character vector with 38 entriey:
\describe{
  \item{names}{name of the variables}

}}
\source{
\url{https://github.com/fabm-model/fabm/wiki/List-of-standard-variables}
}
\usage{
std_names_FABM
}
\description{
A dataset containing the standard names of dependency
variables from physical host models in FABM.
}
\keyword{datasets}
